🚚 move: hooks menu 관련들 query 폴더로 이동(#257)
