+++ src/components/menu-category/container/card/Card.tsx
@@ -1,5 +1,5 @@
 import styles from '@/components/menu-category/styles/category.module.css';
-import useSetCategories from '@/hooks/menu/menu-category/useSetCategories';
+import useSetCategories from '@/hooks/query/menu/menu-category/useSetCategories';
 import useCategoriesStore from '@/shared/store/menu/menu-category';
 import { Tables } from '@/types/supabase';
 import clsx from 'clsx';
